refactor(fatigue): split utils.fatigue into utils/_fatigue package (WP2.4 Phase B)
Behavior-preserving package split modeled on WP2.1 (utils/_profile_estimator/).
The four banner-delimited concerns move VERBATIM into an internal package while
utils/fatigue.py becomes a pure re-export facade — the public surface and import
order stay byte-identical, so every external importer (fatigue_data,
fatigue_context, routes/user_profile, the stage4 observer + calibration scripts,
tests/test_fatigue) keeps working unchanged.

Leaves (leaves-first deps; siblings imported directly, no facade<->leaf cycle):
- utils/_fatigue/core.py       — Phase 1 per-set / session / weekly math + bands.
- utils/_fatigue/per_muscle.py — Phase 2 per-muscle channel (deps: core +
                                 volume_taxonomy).
- utils/_fatigue/period.py     — Phase 2 period selector + logged adapters
                                 (deps: core + per_muscle).
- utils/_fatigue/sfr.py        — Phase 2 Stimulus-to-Fatigue Ratio (no sibling dep).

Mechanical move only: all 42 moved symbol bodies are byte-identical to the
pre-split blob (AST source-segment compare, LF-normalized). No reordering, no
cleanup, and NO consolidation of the scored-row / sort tie-break duplication —
that stays logged for WP2.5 (docs-only). Thresholds, bands, labels, MRV
landmarks, period-window math, and Stage-4 observer artifacts are untouched.

Guarded by the Phase A golden (tests/test_fatigue_golden.py) which stays green
(0 calc drift), plus test_fatigue / _context / _routes / _stage4_observer.
