+++ utils/_fatigue/__init__.py
@@ -1,0 +1,3 @@
+"""Internal staged decomposition of utils.fatigue (Refactor Plan v3 Phase 2,
+WP2.4). Not a public API — import from utils.fatigue instead.
+"""
